Update t4-customs-cassey
Even cheatier method than the cheaty method that the tutor mentioned in class! xD
Just declare a global variable called tallestPrevPerson that stores the height of the tallest person so far in the whole stack, you don't even have to maintain 2 stacks or change the height to be equal to the previous tallest person (the 160 160 160 170 method mentioned in class) 
Is it fool proof? I have no idea though lol